Correct code as per reviewer request
